# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API de Controle de Ponto
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /funcionarios/{id}/registros:
    get:
      operationId: getRegistrosFuncionario
      tags: [registros]
      summary: Folha mensal de um funcionário
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: mes
          in: query
          required: true
          schema:
            type: string
            example: "2025-04"
      responses:
        "200":
          description: Folha mensal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolhaMensal"
  /registros:
    post:
      operationId: upsertRegistro
      tags: [registros]
      summary: Criar ou atualizar registro de ponto de um dia
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertRegistroBody"
      responses:
        "200":
          description: Registro salvo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistroPonto"
  /ponto/bater:
    post:
      operationId: baterPonto
      tags: [registros]
      summary: Registrar entrada ou saída com horário atual
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BaterPontoBody"
      responses:
        "200":
          description: Ponto registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BaterPontoResponse"
...
components:
  schemas:
...
    Funcionario:
      type: object
      properties:
        id:
          type: integer
        codigo:
          type: integer
        nome:
          type: string
        cargo:
          type: string
        vinculo:
          type: string
        situacao:
          type: string
        adiantamento:
          type: number
          format: float
        transporte:
          type: boolean
        vale_alimentacao:
          type: boolean
        jornada_diaria:
          type: string
        ativo:
          type: boolean
        empresa:
          type: string
          nullable: true
        data_contrato:
          type: string
          nullable: true
        salario:
          type: string
          nullable: true
        endereco:
          type: string
          nullable: true
        numero:
          type: string
          nullable: true
        bairro:
          type: string
          nullable: true
        cidade:
          type: string
          nullable: true
        cep:
          type: string
          nullable: true
        estado_civil:
          type: string
          nullable: true
        raca_cor:
          type: string
          nullable: true
        horario:
          type: string
          nullable: true
        escolaridade:
          type: string
          nullable: true
        pis:
          type: string
          nullable: true
        escala_quinzenal:
          type: boolean
        quinzena_referencia:
          type: string
          nullable: true
        he_100_acima_2h:
          type: boolean
        criado_em:
          type: string
      required:
        - id
        - codigo
        - nome
        - cargo
        - vinculo
        - situacao
        - adiantamento
        - transporte
        - jornada_diaria
        - ativo
        - escala_quinzenal
        - he_100_acima_2h
        - criado_em
...
    RegistroPonto:
      type: object
      properties:
        id:
          type: integer
        funcionario_id:
          type: integer
        data:
          type: string
        entrada:
          type: string
          nullable: true
        saida:
          type: string
          nullable: true
        saida_almoco:
          type: string
          nullable: true
        volta_almoco:
          type: string
          nullable: true
        intervalo:
          type: string
          nullable: true
        total_horas:
          type: string
          nullable: true
        he_60:
          type: string
          nullable: true
        he_100:
          type: string
          nullable: true
        atrasos:
          type: string
          nullable: true
        faltas:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        justificativa:
          type: string
          enum: [nenhuma, justificada, injustificada]
        horas_justificadas:
          type: string
          nullable: true
        tipo_dia:
          type: string
          enum: [normal, feriado, feriado_trabalhado, falta, falta_justificada, atraso_justificado]
      required:
        - id
        - funcionario_id
        - data
        - tipo_dia
    UpsertRegistroBody:
      type: object
      properties:
        funcionario_id:
          type: integer
        data:
          type: string
        entrada:
          type: string
          nullable: true
        saida:
          type: string
          nullable: true
        saida_almoco:
          type: string
          nullable: true
        volta_almoco:
          type: string
          nullable: true
        intervalo:
          type: string
          nullable: true
        he_60:
          type: string
          nullable: true
        he_100:
          type: string
          nullable: true
        atrasos:
          type: string
          nullable: true
        faltas:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        justificativa:
          type: string
          enum: [nenhuma, justificada, injustificada]
        tipo_dia:
          type: string
          enum: [normal, feriado, feriado_trabalhado, falta, falta_justificada, atraso_justificado]
      required:
        - funcionario_id
        - data
    BaterPontoBody:
      type: object
      properties:
        funcionario_id:
          type: integer
        tipo:
          type: string
          enum: [entrada, saida_almoco, volta_almoco, saida]
      required:
        - funcionario_id
        - tipo
    BaterPontoResponse:
      type: object
      properties:
        funcionario_id:
          type: integer
        tipo:
          type: string
        horario:
          type: string
        data:
          type: string
        registro:
          $ref: "#/components/schemas/RegistroPonto"
      required:
        - funcionario_id
        - tipo
        - horario
        - data
        - registro
    FolhaMensal:
      type: object
      properties:
        funcionario:
          $ref: "#/components/schemas/Funcionario"
        mes:
          type: string
        registros:
          type: array
          items:
            $ref: "#/components/schemas/RegistroPonto"
        resumo:
          $ref: "#/components/schemas/ResumoMes"
      required:
        - funcionario
        - mes
        - registros
        - resumo
    ResumoMes:
      type: object
      properties:
        total_horas:
          type: string
        he_60:
          type: string
        he_100:
          type: string
        atrasos:
          type: string
        faltas_dia:
          type: number
        faltas_horas:
          type: string
        dias_trabalhados:
          type: integer
        dom_feriados:
          type: integer
        horas_justificadas:
          type: string
        dias_justificados:
          type: integer
      required:
        - total_horas
        - he_60
        - he_100
        - atrasos
        - faltas_dia
        - faltas_horas
        - dias_trabalhados
        - dom_feriados
        - horas_justificadas
        - dias_justificados
...